requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Pillow==10.0.0
Werkzeug==2.3.7
